Refactor ADSR Envelope
